Refactored logic that "smooths" distance between red dots data using fourier transform. The goal of avoiding sudden jumps in distance between red dots is to avoid even larger jumps in zoom_factor -> derivative of distance (change in distance). zoom_factor is needed to adjust drifts from each FeatureMatcher against image zooming in/out due to change in depth.
